extracted out sessionmanager
