+++ aux_ui/lib/screens/link_spotify.dart
@@ -5,13 +5,11 @@ import 'package:flutter/rendering.dart';
 import 'package:aux_ui/widgets/layout/nux_container.dart';
 import 'package:aux_ui/widgets/buttons/icon_bar_button.dart';
 import 'package:spotify_sdk/spotify_sdk.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:logger/logger.dart';
-import 'dart:async';
 
 class LinkSpotify extends SequentialWidget {
-  const LinkSpotify({Key key, String nextPage})
+  final sessionManager;
+
+  const LinkSpotify({Key key, String nextPage, this.sessionManager})
       : super(key: key, nextPage: nextPage);
 
   _LinkSpotifyState createState() => _LinkSpotifyState();
@@ -21,49 +19,10 @@ class _LinkSpotifyState extends State<LinkSpotify> {
   bool _initialized = false;
   Widget _accountSetupText;
   Widget _spotifyLink;
-  final Logger _logger = Logger();
-  var _authToken;
-  bool _connected;
-
-  Future<void> connectToSpotifyRemote() async {
-    try {
-      var result = await SpotifySdk.connectToSpotifyRemote(
-          clientId: DotEnv().env['CLIENT_ID'],
-          redirectUrl: DotEnv().env['REDIRECT_URL']);
-      setState(() {
-        _connected = result;
-      });
-    } on PlatformException catch (e) {
-      setStatus(e.code, message: e.message);
-    } on MissingPluginException {
-      setStatus("not implemented");
-    }
-  }
-
-  Future<void> getAuthenticationToken() async {
-    try {
-      var authenticationToken = await SpotifySdk.getAuthenticationToken(
-          clientId: DotEnv().env['CLIENT_ID'],
-          redirectUrl: DotEnv().env['REDIRECT_URL']);
-      setState(() {
-        _authToken = authenticationToken;
-      });
-    } on PlatformException catch (e) {
-      setStatus(e.code, message: e.message);
-    } on MissingPluginException {
-      setStatus("not implemented");
-    }
-  }
-
-  void setStatus(String code, {String message = ""}) {
-    var text = message.isEmpty ? "" : " : $message";
-    _logger.i("$code$text");
-  }
 
   login() async {
-    await connectToSpotifyRemote();
-    await getAuthenticationToken();
-    if (_connected) widget.next(context); //TODO: subscription connection vs remote connection?
+    bool authConnected = await widget.sessionManager.login();
+    if (authConnected) widget.next(context);
   }
 
   void _initializeWidgets() {
@@ -108,17 +67,18 @@ class _LinkSpotifyState extends State<LinkSpotify> {
     _initializeWidgets();
 
     return StreamBuilder(
-      stream: SpotifySdk.subscribeConnectionStatus(),
+        stream: SpotifySdk.subscribeConnectionStatus(),
         builder: (context, snapshot) {
-        bool _connected = false;
-        if (snapshot.data != null) {
-          _connected = snapshot.data.connected; // TODO: figure out what to do with this/can remove?
-        }
-      return NuxContainer(
-          topFlex: 6,
-          title: 'aux',
-          topWidget: _accountSetupText,
-          bottomWidget: _spotifyLink);
-    });
+//          bool _connected = false;
+//          if (snapshot.data != null) {
+//            _connected = snapshot.data
+//                .connected; // TODO: figure out what to do with this/can remove streambuilder?
+//          }
+          return NuxContainer(
+              topFlex: 6,
+              title: 'aux',
+              topWidget: _accountSetupText,
+              bottomWidget: _spotifyLink);
+        });
   }
 }
